Starting to add some tests and extracting some utils
