+++ CONTRIBUTING.md
@@ -1,0 +1,16 @@
+# Contributing to teams
+
+Thanks for considering a contribution! All contributions are welcome, including
+bug reports, bug fixes, new features or even just questions.
+
+### Features
+
+For PRs, please:
+- add new tests to cover the new code
+- make sure all of the tests pass (`make test`)
+
+### Bugs
+
+Submit issues to [issue tracker](https://github.com/hairyhenderson/teams/issues/).
+
+If you think you've found a sensitive security issue, please e-mail me before opening an issue: dhenderson@gmail.com. My PGP key is available on Keybase: https://keybase.io/dhenderson/.
