+++ circle.yml
@@ -1,0 +1,7 @@
+test:
+  pre:
+    - go get -u github.com/jstemmer/go-junit-report
+  override:
+    - |
+      set -e -o pipefail
+      go test -v -race ./... | go-junit-report > $CIRCLE_TEST_REPORTS/report.xml
